requirements.txt
fastapi
uvicorn
python-dotenv
google-adk
openai
pydantic
